put <ctrl-a> back in, just seems too windows-ey not to have. Changed it to go into visual mode instead of selection mode so you can use it to apply commands to the entire file instead of selection mode which just deletes the selection and starts typing like a regular win editor.
Mapped old behavior of <ctrl-a> to <F12>.

Made <ctrl-c> maintain the visual selection after the yank. It's more
windows-ey that way and gives a nice alternative to 'y'.

Also discovered that <A-Space> does work if you remap it after vim is
loaded. I think this means <a-space> is getting remapped by a plugin.
Not too worried about it for now but it might be nice to get to the bottom
of that.